Fixed bugs with mixed up long/long long

+++ 8cc.h
@@ -141,7 +141,7 @@ typedef struct Node {
     SourceLoc *sourceLoc;
     union {
         // Char, int, or long
-        long ival;
+        long long ival;
         // Float or double
         struct {
             double fval;
